Seguridad
Agrega "seguridad" para que los usuarios no accedan a información de otros usuarios

+++ Todo/todo.py
@@ -12,7 +12,9 @@ bp = Blueprint("todo", __name__)
 def index():
     db, c = get_db()
     c.execute(
-        'SELECT t.id, t.description, u.username, t.completed, t.created_at FROM todo t JOIN user u ON t.created_by=u.id ORDER BY created_at DESC'
+        'SELECT t.id, t.description, u.username, t.completed, t.created_at '
+        'FROM todo t JOIN user u ON t.created_by=u.id WHERE t.created_by = %s ORDER BY created_at DESC',
+        (g.user['id'],)
     )
     todos = c.fetchall()
     return render_template("todo/index.html", todos=todos)
@@ -70,8 +72,8 @@ def update(id):
             db, c = get_db()
             c.execute(
                 "UPDATE todo SET description = %s, completed = %s "
-                "WHERE id = %s",
-                (description, completed, id)
+                "WHERE id = %s AND created_by = %s",
+                (description, completed, id, g.user['id'])
             )
             db.commit()
             return redirect(url_for("todo.index"))
@@ -82,7 +84,7 @@ def update(id):
 @login_required
 def delete(id):
     db, c = get_db()
-    c.execute("DELETE FROM todo WHERE id=%s", (id,))
+    c.execute("DELETE FROM todo WHERE id=%s AND created_by = %s", (id, g.user['id']))
     db.commit()
     return redirect(url_for("todo.index"))
